requirements.txt
Django==1.8
gunicorn
psycopg2
dj-database-url
djangorestframework
python-google-places==0.12.1
